requirements.txt
requests
notebook
Pillow
black
openai
